backend feedback
upadted

+++ backend/src/main/java/com/example/backend/Repo/FeedBackRepo.java
@@ -3,5 +3,5 @@ package com.example.backend.Repo;
 import com.example.backend.entity.FeedBack;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedBackRepo extends JpaRepository<FeedBack,Long> {
+public interface FeedBackRepo extends JpaRepository<FeedBack,Integer> {
 }
